requirements.txt
grpcio==1.58.0
grpcio-tools==1.58.0
protobuf==4.24.3
pytest==7.4.2
python-dotenv==1.0.0
